+++ packages/remax-plugin-error-screen/src/index.ts
@@ -1,6 +1,7 @@
 import * as path from 'path';
 import * as fs from 'fs';
 import VirtualModulesPlugin from 'webpack-virtual-modules';
+import { slash } from '@remax/shared';
 
 export default (_: any, { cwd, rootDir }: { cwd: string; rootDir: string }) => {
   const searchCustomErrorFile = () => {
@@ -26,18 +27,20 @@ export default (_: any, { cwd, rootDir }: { cwd: string; rootDir: string }) => {
     'node_modules/@remax/plugin-error-screen/runtime.js': `
         import React from 'react';
         import { View } from 'remax/one';
-        import ErrorScreen from '${errorScreenFile}';
-        import ErrorBoundary from '${errorBoundaryFile}';
+        import ErrorScreen from '${slash(errorScreenFile)}';
+        import ErrorBoundary from '${slash(errorBoundaryFile)}';
 
         export default {
           onPageComponent({ component }) {
-            return React.forwardRef((props, ref) => {
+            function ErrorBoundaryWrap(props, ref) {
               return React.createElement(
                 ErrorBoundary,
                 { errorScreen: ErrorScreen },
                 React.createElement(component, { ...props, ref })
               );
-            });
+            }
+            ErrorBoundaryWrap.displayName = 'ErrorBoundary'
+            return React.forwardRef(ErrorBoundaryWrap);
           },
         };
       `,