+++ packages/remax-web/src/hooks/usePageInstance.ts
@@ -1,0 +1,3 @@
+export default function usePageInstance(): any {
+  return null;
+}